Dnese nodes iterator

+++ src/main/scala/com/acervera/pbf4scala/model/NodeEntity.scala
@@ -3,8 +3,7 @@ package com.acervera.pbf4scala.model
 /**
   * Created by angelcervera on 14/06/16.
   */
-case class NodeEntity(val id: Long, val latitude: Long, val longitude: Long, val tags: Map[String, String]) extends OSMEntity {
-
+case class NodeEntity(val id: Long, val latitude: Double, val longitude: Double, val tags: Map[String, String]) extends OSMEntity {
   override val osmModel: OSMTypes.Value = OSMTypes.Node
-
 }
+
